requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
faiss-cpu
langchain-huggingface
groq
langchain_groq
-e .